+++ sidebars-community.ts
@@ -1,8 +1,5 @@
-// @ts-check
-
-/** @type {import('@docusaurus/plugin-content-docs').SidebarsConfig} */
-
-const sidebars = {
+import type {SidebarsConfig} from '@docusaurus/plugin-content-docs';
+const sidebars: SidebarsConfig = {
   kb: [
     // Docs Link from Community
     {
@@ -52,4 +49,5 @@ const sidebars = {
   ],
 };
 
-module.exports = sidebars;
+
+export default sidebars;
