+++ sidebars-kb.ts
@@ -1,8 +1,5 @@
-// @ts-check
-
-/** @type {import('@docusaurus/plugin-content-docs').SidebarsConfig} */
-
-const sidebars = {
+import type {SidebarsConfig} from '@docusaurus/plugin-content-docs';
+const sidebars: SidebarsConfig = {
   kb: [
     // KB Parent
     {
@@ -333,4 +330,4 @@ const sidebars = {
   ],
 };
 
-module.exports = sidebars;
+export default sidebars;
