+++ sidebars.ts
@@ -1,8 +1,5 @@
-// @ts-check
-
-/** @type {import('@docusaurus/plugin-content-docs').SidebarsConfig} */
-
-const sidebars = {
+import type {SidebarsConfig} from '@docusaurus/plugin-content-docs';
+const sidebars: SidebarsConfig = {
   allcontent: [
     // Documentation Parent
     {
@@ -1461,4 +1458,4 @@ const sidebars = {
   ],
 };
 
-module.exports = sidebars;
+export default sidebars;
